pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "churn-ml-decision"
version = "0.1.0"
description = "Churn prediction and business decision pipeline."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pandas",
  "scikit-learn",
  "matplotlib",
  "seaborn",
  "joblib",
  "scipy",
  "pyyaml",
  "pydantic>=2,<3",
  "tenacity",
]

[project.optional-dependencies]
ml = ["xgboost", "lightgbm", "shap"]
ops = ["dvc", "mlflow"]
dev = ["pytest", "pytest-cov", "ruff"]

[project.scripts]
churn-prepare = "churn_ml_decision.prepare:main"
churn-train = "churn_ml_decision.train:main"
churn-evaluate = "churn_ml_decision.evaluate:main"
churn-predict = "churn_ml_decision.predict:main"
churn-validate-config = "churn_ml_decision.cli:validate_config_main"
churn-model-info = "churn_ml_decision.cli:model_info_main"
churn-model-promote = "churn_ml_decision.cli:model_promote_main"
churn-model-rollback = "churn_ml_decision.cli:model_rollback_main"
churn-check-drift = "churn_ml_decision.cli:check_drift_main"
churn-health-check = "churn_ml_decision.cli:health_check_main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
